product score

+++ coastal-web/coastal/apps/product/models.py
@@ -145,7 +145,7 @@ class Product(models.Model):
     desc_other_to_note = models.TextField(max_length=255, null=True, blank=True)
 
     # score
-    score = models.PositiveSmallIntegerField(default=0)
+    score = models.PositiveIntegerField(default=0)
 
     @property
     def short_desc(self):
